--- a/apps/ssr-tests-v9/jest.config.js
+++ b/apps/ssr-tests-v9/jest.config.js
@@ -7,13 +7,7 @@ module.exports = {
   displayName: 'ssr-tests-v9',
   preset: '../../jest.preset.js',
   transform: {
-    '^.+\\.tsx?$': [
-      'ts-jest',
-      {
-        tsconfig: '<rootDir>/tsconfig.spec.json',
-        isolatedModules: true,
-      },
-    ],
+    '^.+\\.tsx?$': ['@swc/jest', {}],
   },
   testEnvironment: 'node',
   coverageDirectory: './coverage',