--- a/apps/vr-tests-react-components/jest.config.js
+++ b/apps/vr-tests-react-components/jest.config.js
@@ -7,12 +7,6 @@ module.exports = {
   displayName: 'vr-tests-react-components',
   preset: '../../jest.preset.js',
   transform: {
-    '^.+\\.tsx?$': [
-      'ts-jest',
-      {
-        tsconfig: '<rootDir>/tsconfig.json',
-        isolatedModules: true,
-      },
-    ],
+    '^.+\\.tsx?$': ['@swc/jest', {}],
   },
 };