<!--
Thank you for submitting a pull request!

Please verify that:
* [ ] Code is up-to-date with the `master` branch
* [ ] Your changes are covered by tests (if possible)
* [ ] You've run `yarn change` locally


PR flow tips:
* [ ] Try to start with a Draft PR
* [ ] Once you're ready (ideally the pipeline is passing) promote your PR to Ready for Review. This step will auto-assign reviewers for your PR.
-->

## Previous Behavior

<!-- This is the behavior we have today -->

## New Behavior

> **NOTE:** all the migrations have been done with nx generator and AST transforms ( local machine only )

- unifies transpilation tooling for production and testing in all v9 projects
- fixes mocks that were failing after migration
- bumps `dedent` which was using invalid type definitions that forced us to use invalid ESM imports -> exposed thanks to swc

**Perf**

| command - `nx run-many -t test -p tag:vNext`  | time (ts-jest) | time (swc-jest) |
| ------------------------------ | -------------- | --------------- |
| current (default settings)     | 7m 15s         | 5m 20s  / 𝚫 26%  FASTER       |
| runInBand (maxWorkers=1) / [PR](https://github.com/microsoft/fluentui/pull/32096)       | 4m 25s       | 4m 20s  / 𝚫 2% FASTER     |

| command - `nx run-many -t test -p 'scripts/*'`  | time (ts-jest) | time (swc-jest) |
| ------------------------------ | -------------- | --------------- |
| current     |  18s         | 15s  / 𝚫 16% FASTER         |

## Related Issue(s)

<!-- Please link the issue being fixed so it gets closed when this is merged. -->

- Follows https://github.com/microsoft/fluentui/pull/32096